requirements.txt
requests==2.23.0
Pillow==7.0.0
instabot==0.116.0
python-dotenv==0.12.0